AIML-554: Extract ADK asyncio workarounds from main.py (Phase 8)
Move ~160 lines of asyncio monkey-patching out of main.py into
asyncio_workarounds.py. main.py now calls apply_asyncio_workarounds()
at startup and cleanup_event_loop() at exit — zero behavior change.
