+++ src/main.py
@@ -19,12 +19,7 @@
 
 import sys
 import re
-import asyncio
-import warnings
-import atexit
-import platform
 from datetime import datetime, timedelta
-from asyncio.proactor_events import _ProactorBasePipeTransport
 from urllib.parse import urlparse
 
 # Import configurations and utilities
@@ -47,6 +42,7 @@ from src.smartfix.domains.vulnerability.context import RemediationContext, Promp
 from src.smartfix.domains.vulnerability.models import Vulnerability
 
 from src.smartfix.domains.agents.smartfix_agent import SmartFixAgent
+from src.smartfix.domains.agents.asyncio_workarounds import apply_asyncio_workarounds, cleanup_event_loop
 from src.github.external_coding_agent import ExternalCodingAgent
 from src.smartfix.domains.workflow.pr_reconciliation import reconcile_open_remediations
 
@@ -57,188 +53,7 @@ telemetry_handler.initialize_telemetry()
 git_ops = GitOperations()
 github_ops = GitHubOperations()
 
-# NOTE: Google ADK appears to have issues with asyncio event loop cleanup, and has had attempts to address them in versions 1.4.0-1.5.0
-# Configure warnings to ignore asyncio ResourceWarnings during shutdown
-warnings.filterwarnings("ignore", category=ResourceWarning,
-                        message="unclosed.*<asyncio.sslproto._SSLProtocolTransport.*")
-warnings.filterwarnings("ignore", category=ResourceWarning,
-                        message="unclosed transport")
-warnings.filterwarnings("ignore", category=ResourceWarning,
-                        message="unclosed.*<asyncio.*")
-
-# Patch asyncio to handle event loop closed errors during shutdown
-_original_loop_check_closed = asyncio.base_events.BaseEventLoop._check_closed
-
-
-def _patched_loop_check_closed(self):
-    try:
-        _original_loop_check_closed(self)
-    except RuntimeError as e:
-        if "Event loop is closed" in str(e):
-            return  # Suppress the error
-        raise
-
-
-asyncio.BaseEventLoop._check_closed = _patched_loop_check_closed
-
-
-# Add a specific fix for _ProactorBasePipeTransport.__del__ on Windows
-if platform.system() == 'Windows':
-    # Import the specific module that contains ProactorBasePipeTransport
-    try:
-        # Store the original __del__ method
-        _original_pipe_del = _ProactorBasePipeTransport.__del__
-
-        # Define a safe replacement for __del__
-        def _patched_pipe_del(self):
-            try:
-                # Check if the event loop is closed or finalizing
-                if self._loop.is_closed() or sys.is_finalizing():
-                    # Skip the original __del__ which would trigger the error
-                    return
-
-                # Otherwise use the original __del__ implementation
-                _original_pipe_del(self)
-            except (AttributeError, RuntimeError, ImportError, TypeError):
-                # Catch and ignore all attribute or runtime errors during shutdown
-                pass
-
-        # Apply the patch to the __del__ method
-        _ProactorBasePipeTransport.__del__ = _patched_pipe_del
-
-        debug_log("Successfully patched _ProactorBasePipeTransport.__del__ for Windows")
-    except (ImportError, AttributeError) as e:
-        debug_log(f"Could not patch _ProactorBasePipeTransport: {str(e)}")
-
-    # Add a specific fix for BaseSubprocessTransport.__del__ on Windows
-    try:
-        from asyncio.base_subprocess import BaseSubprocessTransport
-
-        # Store the original __del__ method
-        _original_subprocess_del = BaseSubprocessTransport.__del__
-
-        # Define a safe replacement for __del__
-        def _patched_subprocess_del(self):
-            try:
-                # Check if the event loop is closed or finalizing
-                if hasattr(self, '_loop') and self._loop is not None and (self._loop.is_closed() or sys.is_finalizing()):
-                    # Skip the original __del__ which would trigger the error
-                    return
-
-                # Otherwise use the original __del__ implementation
-                _original_subprocess_del(self)
-            except (AttributeError, RuntimeError, ImportError, TypeError, ValueError):
-                # Catch and ignore all attribute, runtime, or value errors during shutdown
-                # ValueError specifically handles "I/O operation on closed pipe"
-                pass
-
-        # Apply the patch to the __del__ method
-        BaseSubprocessTransport.__del__ = _patched_subprocess_del
-
-        debug_log("Successfully patched BaseSubprocessTransport.__del__ for Windows")
-    except (ImportError, AttributeError) as e:
-        debug_log(f"Could not patch BaseSubprocessTransport: {str(e)}")
-
-
-def cleanup_asyncio():  # noqa: C901
-    """
-    Cleanup function registered with atexit to properly handle asyncio resources during shutdown.
-    This helps prevent the "Event loop is closed" errors during program exit.
-    """
-    # Suppress stderr temporarily to avoid printing shutdown errors
-    original_stderr = sys.stderr
-    try:
-        # Create a dummy stderr to suppress errors during cleanup
-        class DummyStderr:
-            def write(self, *args, **kwargs):
-                pass
-
-            def flush(self):
-                pass
-
-        # Only on Windows do we need the more aggressive error suppression
-        if platform.system() == 'Windows':
-            sys.stderr = DummyStderr()
-
-            # Windows-specific: ensure the proactor event loop resources are properly cleaned
-            try:
-                # Try to access the global WindowsProactorEventLoopPolicy
-                loop_policy = asyncio.get_event_loop_policy()
-
-                # If we have any running loops, close them properly
-                try:
-                    loop = loop_policy.get_event_loop()
-                    if not loop.is_closed():
-                        if loop.is_running():
-                            loop.stop()
-
-                        # Cancel all tasks
-                        pending = asyncio.all_tasks(loop)
-                        if pending:
-                            for task in pending:
-                                task.cancel()
-
-                            # Give tasks a chance to respond to cancellation with a timeout
-                            try:
-                                loop.run_until_complete(asyncio.wait_for(
-                                    asyncio.gather(*pending, return_exceptions=True),
-                                    timeout=1.0
-                                ))
-                            except (asyncio.CancelledError, asyncio.TimeoutError, Exception):
-                                pass
-
-                        # Close transports and other resources
-                        try:
-                            loop.run_until_complete(loop.shutdown_asyncgens())
-                        except Exception:
-                            pass
-
-                        try:
-                            loop.close()
-                        except Exception:
-                            pass
-                except Exception:
-                    pass
-
-                # Force garbage collection to ensure __del__ methods are called
-                try:
-                    import gc
-                    gc.collect()
-                except Exception:
-                    pass
-
-            except Exception:
-                pass  # Ignore any errors during Windows-specific cleanup
-        else:
-            # For non-Windows platforms, perform regular cleanup
-            try:
-                loop = asyncio.get_event_loop()
-                if loop.is_running():
-                    loop.stop()
-
-                # Cancel all tasks
-                pending = asyncio.all_tasks(loop)
-                if pending:
-                    for task in pending:
-                        task.cancel()
-
-                    # Give tasks a chance to respond to cancellation
-                    if not loop.is_closed():
-                        loop.run_until_complete(asyncio.gather(*pending, return_exceptions=True))
-
-                # Close the loop
-                if not loop.is_closed():
-                    loop.run_until_complete(loop.shutdown_asyncgens())
-                    loop.close()
-            except Exception:
-                pass  # Ignore any errors during cleanup
-    finally:
-        # Restore stderr
-        sys.stderr = original_stderr
-
-
-# Register the cleanup function
-atexit.register(cleanup_asyncio)
+apply_asyncio_workarounds()
 
 
 def main():  # noqa: C901
@@ -741,50 +556,7 @@ def main():  # noqa: C901
 
     log(f"\n--- Script finished (total runtime: {total_runtime}) ---")
 
-    # Clean up any dangling asyncio resources
-    try:
-        # Force asyncio resource cleanup before exit
-        loop = asyncio.get_event_loop_policy().get_event_loop()
-        if not loop.is_closed():
-            # Cancel all pending tasks
-            pending = asyncio.all_tasks(loop)
-            if pending:
-                for task in pending:
-                    try:
-                        task.cancel()
-                    except Exception:
-                        pass
-
-                # Give tasks a chance to respond to cancellation
-                try:
-                    # Wait with a timeout to prevent hanging
-                    loop.run_until_complete(asyncio.gather(*pending, return_exceptions=True))
-                except (asyncio.CancelledError, Exception):
-                    pass
-
-            try:
-                # Shut down asyncgens
-                loop.run_until_complete(loop.shutdown_asyncgens())
-            except Exception:
-                pass
-
-            try:
-                # Close the loop
-                loop.close()
-            except Exception:
-                pass
-
-        # On Windows, specifically force garbage collection
-        if platform.system() == 'Windows':
-            try:
-                import gc
-                gc.collect()
-            except Exception:
-                pass
-    except Exception as e:
-        # Ignore any errors during cleanup
-        debug_log(f"Ignoring error during asyncio cleanup: {str(e)}")
-        pass
+    cleanup_event_loop()
 
 
 if __name__ == "__main__":
